+++ app/components/layout/index.ts
@@ -6,3 +6,4 @@ export * from "./model-select";
 export * from "./titlebar";
 export * from "./mobile-bottom-nav";
 export * from "./mobile-play-header";
+export * from "./whats-new-modal";